moved expandable transaction to mint folder since not generic component
